interp: rewrite argument processing
Ultimately this bug was caused by:

```
if len(unnamedArgs)-len(formalParams) > 0 {
```

Should have been:

```
if len(unnamedArgs)-len(paramData.Params) > 0 {
```

`formalParams` included `paramData.KWParams`, which should not have been
considered for excess arg handling.

I believe that this is overall a better implementation however, since I
needed to rewrite it to understand and debug it properly, even though it
was a fairly trivial bug.

So the rewrite stays, instead of being a small bug fix.