interp/flow: refactor to differentiate between globals and intrinsics
This is a thing where lambdas have access to things such as `Args`,
`min`, `max`, etc, but not `data`, `const`, etc.

It's not clear what's intrinsic and what's global yet, this just sets
up the framework.

Our scope rules change such that at the bottom layer we have Intrinsics,
then we have Globals on top of that.  Then functions are on top of that.

When lambdas come in, they will be given access to Intrindics, and their
own scope with the Identifier, but nothing else.